@@ -20,11 +20,12 @@
   },
   "scripts": {
     "build": "turbo run build --filter=./packages/*",
-    "build:examples": "turbo run build --filter=./examples/*",
     "clean": "turbo run clean",
-    "format": "prettier --write . && bun run format:pack",
-    "format:pack": "npx sort-package-json package.json packages/**/**/package.json examples/*/package.json docs/package.json e2e/package.json",
-    "generate": "turbo run generate --continue && bun run format",
+    "format": "bun run format:dprint",
+    "format:dprint": "dprint fmt",
+    "format:fix": "dprint fmt --incremental=false",
+    "format:pack": "npx sort-package-json package.json packages/**/**/package.json",
+    "format:prettier": "prettier --config ./configs/prettier.config.cjs --write .",
     "graph": "npx nx graph ",
     "install:csb": "corepack enable && pnpm install --frozen-lockfile",
     "lint": "ESLINT_USE_FLAT_CONFIG=true eslint --format pretty ./packages/*",
@@ -32,36 +33,36 @@
     "lint:ci": "ESLINT_USE_FLAT_CONFIG=true eslint --max-warnings 10 --format pretty ./packages/*",
     "lint:fix": "ESLINT_USE_FLAT_CONFIG=true eslint --fix --format pretty ./packages/* && bun run lint:case",
     "lint:turbo": "turbo run lint",
-    "release": "turbo run release --filter=./packages/* --continue",
-    "release:canary": "turbo run release:canary --filter=./packages/* --continue",
+    "release": "changeset version && changeset publish",
+    "release:canary": "changeset version --snapshot && changeset publish",
     "start": "turbo run start --filter=./packages/*",
     "test": "vitest run --config ./configs/vitest.config.ts --coverage",
     "test:bun": "bun test --preload ./configs/setup.ts",
-    "test:jest": "npx jest --config ./configs/jest.config.ts",
     "test:vitest": "vitest run --config ./configs/vitest.config.ts",
     "test:watch": "vitest --config ./configs/vitest.config.ts",
-    "typecheck": "turbo run typecheck --continue --filter='./packages/*' --filter='./examples/*'",
-    "upgrade": "npx taze -r -w --exclude pnpm"
+    "typecheck": "turbo run typecheck --continue --filter='./packages/*'",
+    "upgrade": "npx taze -r -w --exclude pnpm",
+    "changeset": "changeset",
+    "//changeset:beta:enter": "changeset pre enter beta",
+    "//changeset:beta:exit": "changeset pre exit"
   },
   "devDependencies": {
-    "@ianvs/prettier-plugin-sort-imports": "^4.1.0",
+    "@changesets/changelog-github": "^0.4.8",
+    "@changesets/cli": "^2.26.2",
     "@stijnvanhulle/eslint-config": "workspace:*",
     "@stijnvanhulle/ts-config": "workspace:*",
-    "@swc/core": "^1.3.76",
-    "@swc/jest": "^0.2.28",
-    "@types/node": "^20.4.9",
-    "@vitest/coverage-v8": "^0.34.1",
-    "@vitest/ui": "^0.34.1",
-    "bun-types": "^0.7.3",
-    "jest": "^29.6.2",
-    "prettier": "^3.0.1",
-    "prettier-eslint": "^15.0.1",
-    "prettier-plugin-curly": "^0.1.2",
-    "rimraf": "^5.0.1",
+    "@types/node": "^20.10.0",
+    "@vitest/coverage-v8": "^0.34.6",
+    "@vitest/ui": "^0.34.6",
+    "bun-types": "^1.0.14",
+    "dprint": "^0.41.0",
+    "prettier": "^3.1.0",
+    "prettier-eslint": "^16.1.2",
+    "rimraf": "^5.0.5",
     "ts-node": "^10.9.1",
-    "turbo": "^1.10.12",
-    "typescript": "~5.1.6",
-    "vitest": "^0.34.1"
+    "turbo": "^1.10.16",
+    "typescript": "^5.3.2",
+    "vitest": "^0.34.6"
   },
   "packageManager": "pnpm@8.3.0",
   "engines": {
