@@ -1,74 +1,71 @@
 {
   "name": "@stijnvanhulle/eslint-config",
-  "version": "2.1.0",
+  "version": "1.1.8",
   "private": true,
   "license": "MIT",
   "type": "module",
   "exports": {
     ".": {
       "import": "./dist/index.js",
-      "module": "./dist/index.js",
+      "require": "./dist/index.cjs",
       "default": "./dist/index.cjs"
+    },
+    "./legacy": {
+      "import": "./dist/legacy.js",
+      "require": "./dist/legacy.cjs",
+      "default": "./dist/legacy.cjs"
     },
     "./flat": {
       "import": "./dist/flat.js",
-      "module": "./dist/flat.js",
+      "require": "./dist/flat.cjs",
       "default": "./dist/flat.cjs"
     },
     "./package.json": "./package.json",
     "./*": "./*"
   },
-  "main": "dist/index.js",
+  "main": "dist/index.cjs",
   "module": "dist/index.js",
   "types": "./dist/index.d.ts",
-  "typesVersions": {
-    "*": {
-      "index": [
-        "./dist/index.d.ts"
-      ],
-      "flat": [
-        "./dist/flat.d.ts"
-      ]
-    }
-  },
   "files": [
+    "src",
     "dist",
-    "schemas",
     "!/**/**.test.**",
     "!/**/__tests__/**"
   ],
   "scripts": {
     "build": "tsup",
-    "clean": "rimraf ./dist",
+    "clean": "npx rimraf ./dist",
     "lint": "ESLINT_USE_FLAT_CONFIG=true eslint --format pretty .",
     "lint-fix": "bun run lint --quiet --fix",
     "release": "pnpm publish --no-git-check",
-    "release:canary": "bash ../../../.github/canary.sh && pnpm publish --no-git-check -tag canary",
+    "release:canary": "bash ../../../.github/canary.sh && node ../../../scripts/build.js canary && pnpm publish --no-git-check",
     "start": "tsup --watch",
     "typecheck": "tsc -p ./tsconfig.json --noEmit --emitDeclarationOnly false"
   },
   "dependencies": {
-    "@eslint/js": "^8.46.0",
-    "@types/eslint": "~8.44.2",
-    "@typescript-eslint/eslint-plugin": "^6.3.0",
-    "@typescript-eslint/parser": "~6.3.0",
+    "@eslint/js": "^8.54.0",
+    "@types/eslint": "~8.44.7",
+    "@typescript-eslint/eslint-plugin": "^6.12.0",
+    "@typescript-eslint/parser": "~6.7.5",
     "eslint-config-prettier": "~9.0.0",
-    "eslint-config-turbo": "~1.10.12",
+    "eslint-config-turbo": "~1.10.16",
     "eslint-formatter-pretty": "^5.0.0",
     "eslint-plugin-eslint-plugin": "^5.1.1",
-    "eslint-plugin-import": "~2.28.0",
-    "eslint-plugin-react": "~7.33.1",
+    "eslint-plugin-import": "~2.28.1",
+    "eslint-plugin-react": "~7.33.2",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint-plugin-turbo": "~1.10.16",
     "eslint-plugin-unused-imports": "~3.0.0",
-    "eslint-plugin-vitest": "~0.2.8",
+    "eslint-plugin-vitest": "~0.3.10",
     "eslint-plugin-vitest-globals": "~1.4.0",
-    "globals": "^13.20.0"
+    "globals": "^13.23.0"
   },
   "devDependencies": {
     "@stijnvanhulle/ts-config": "workspace:*",
     "@stijnvanhulle/tsup-config": "workspace:*",
-    "eslint": "~8.46.0",
-    "tsup": "^7.2.0",
-    "typescript": "~5.1.6"
+    "eslint": "^8.54.0",
+    "tsup": "^8.0.1",
+    "typescript": "~5.2.2"
   },
   "packageManager": "pnpm@8.3.0",
   "engines": {
