@@ -2,11 +2,9 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 // prettier-ignore-start
-
 /**
  * Configs
  */
-
 // @ts-ignore
 import eslint from '@eslint/js'
 import tsPlugin from '@typescript-eslint/eslint-plugin'
@@ -17,15 +15,17 @@ import prettier from 'eslint-config-prettier'
 /**
  * Recommended plugins ESM
  */
-
 // @ts-ignore
-import eslintPluginRecommended from 'eslint-plugin-eslint-plugin/configs/recommended'
+import eslintPluginRecommended from 'eslint-plugin-eslint-plugin'
 /**
  * Plugins
  */
-
 // @ts-ignore
 import importPlugin from 'eslint-plugin-import'
+// @ts-ignore
+import reactPlugin from 'eslint-plugin-react'
+// @ts-ignore
+import simpleImportSortPlugin from 'eslint-plugin-simple-import-sort'
 // @ts-ignore
 import turboPlugin from 'eslint-plugin-turbo'
 // @ts-ignore
@@ -38,42 +38,39 @@ import vitestGlobalsPlugin from 'eslint-plugin-vitest-globals'
  */
 import globals from 'globals'
 
-import { rules } from './rules'
+import packageJSON from '../package.json'
+import { ignores } from './ignores.ts'
+import { rules } from './rules.ts'
 
-import type { Linter } from 'eslint'
-import { ignores } from './ignores'
-
-/**
- * Recommended plugins CJS
- */
-
-// @ts-ignore
-const reactPluginRecommended = require('eslint-plugin-react/configs/recommended')
+import type { ESLint, Linter } from 'eslint'
 
 // prettier-ignore-end
 
-export const config: Linter.FlatConfig = {
+const recommended = {
   files: ['**/*.{ts,tsx}'],
   ignores: ignores.all,
   rules: {
     ...eslint.configs.recommended.rules,
     ...turboPlugin.configs['recommended'].rules,
-    ...prettier.rules,
     // @ts-ignore
     ...tsPlugin.configs['eslint-recommended'].overrides[0].rules,
-    ...tsPlugin.configs['recommended'].rules,
-    ...tsPlugin.configs['recommended-requiring-type-checking'].rules,
+    ...tsPlugin.configs['recommended']?.rules,
+    ...tsPlugin.configs['recommended-requiring-type-checking']?.rules,
+    ...reactPlugin.configs['recommended']?.rules,
+    ...prettier.rules,
     ...rules,
   },
   plugins: {
-    // @ts-ignore
+    '': eslintPluginRecommended,
+    'react': reactPlugin,
     '@typescript-eslint': tsPlugin,
     vitest: vitestPlugin,
     'vitest-globals': vitestGlobalsPlugin,
     'unused-imports': unusedImportsPlugin,
     import: importPlugin,
     turbo: turboPlugin,
-  },
+    'simple-import-sort': simpleImportSortPlugin,
+  } as unknown as Linter.FlatConfig['plugins'],
   languageOptions: {
     ecmaVersion: 'latest',
     globals: {
@@ -94,7 +91,7 @@ export const config: Linter.FlatConfig = {
        */
       warnOnUnsupportedTypeScriptVersion: false,
     },
-  },
+  } as Linter.FlatConfig['languageOptions'],
   linterOptions: {
     reportUnusedDisableDirectives: true,
   },
@@ -108,57 +105,45 @@ export const config: Linter.FlatConfig = {
     'import/resolver': {
       node: true,
       typescript: {
-        project: `'packages/*/tsconfig.json',`,
+        project: true,
       },
     },
     react: {
       version: 'detect',
     },
-    next: {
-      rootDir: ['docs/*/'],
-    },
   },
-}
+} satisfies Linter.FlatConfig
 
-export const configExamples: Linter.FlatConfig = {
-  files: ['examples/**', 'e2e/**'],
-  ignores: ignores.all,
-  rules: {
-    'import/extensions': [
-      'off',
-      'ignorePackages',
-      {
-        js: 'always',
-        jsx: 'always',
-        ts: 'always',
-        tsx: 'always',
-      },
-    ],
-  },
-}
-
-export const configTest: Linter.FlatConfig = {
+const tests = {
   files: ['**/*.test.ts', '**/*.test.tsx'],
   ignores: ignores.all,
+  plugins: {
+    '@typescript-eslint': tsPlugin,
+  } as unknown as Linter.FlatConfig['plugins'],
   rules: {
     '@typescript-eslint/no-unsafe-assignment': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
     '@typescript-eslint/no-unsafe-member-access': 'off',
     '@typescript-eslint/no-unsafe-call': 'off',
   },
-}
+} satisfies Linter.FlatConfig
 
-export const configDist: Linter.FlatConfig = {
+const build = {
   ignores: ignores.build,
-}
+} satisfies Linter.FlatConfig
 
-export const configs: Linter.FlatConfig[] = [
-  reactPluginRecommended as Linter.FlatConfig,
-  eslintPluginRecommended as Linter.FlatConfig,
-  config,
-  configExamples,
-  configTest,
-  configDist,
-]
+const plugin = {
+  meta: {
+    name: '@stijnvanhulle/eslint-config',
+    version: packageJSON.version,
+  },
+  configs: {
+    recommended,
+    tests,
+    build,
+  },
+  rules: {},
+  processors: {},
+} satisfies ESLint.Plugin
 
-export default config
+export default plugin
