@@ -1,4 +1,4 @@
-export const defaultIgnores = ['dist/**', 'coverage/**', 'mocks/**', '*.d.ts']
+export const defaultIgnores = ['dist/**', 'coverage/**', 'mocks/**', '*.d.ts', '*.config.ts', 'templates/**']
 
 /**
  * When using in a monorepo(with a `packages` folder) we need to append `**\/`
@@ -21,5 +21,3 @@ export const ignores = {
   all: [...defaultIgnores, ...monorepoIgnores],
   build: getMonorepoIgnores(['dist/**', 'coverage/**']),
 }
-
-export default ignores
