@@ -1,10 +1,20 @@
 import { defineConfig } from 'tsup'
 
-import { optionsCJS, optionsESM } from '../tsup-config/src'
+import { optionsCJS, optionsESM } from '../tsup-config/src/index.ts'
 
 export default defineConfig([
   optionsCJS,
   optionsESM,
+  {
+    ...optionsCJS,
+    entry: ['src/legacy.ts'],
+    name: 'legacy',
+  },
+  {
+    ...optionsESM,
+    entry: ['src/legacy.ts'],
+    name: 'legacy',
+  },
   {
     ...optionsCJS,
     entry: ['src/flat.ts'],
