@@ -3,28 +3,25 @@
 
 /**
  * Copy of `@total-typescript/ts-reset`
- * @link 
+ * @link
  */
-type NonFalsy<T> = T extends false | 0 | "" | null | undefined | 0n
-? never
-: T;
+type NonFalsy<T> = T extends false | 0 | '' | null | undefined | 0n ? never : T
 
 /**
  * Copy of `@total-typescript/ts-reset`
- * @link 
+ * @link
  */
 interface Array<T> {
-  filter(predicate: BooleanConstructor, thisArg?: any): NonFalsy<T>[];
+  filter(predicate: BooleanConstructor, thisArg?: any): NonFalsy<T>[]
 }
 
 /**
  * Copy of `@total-typescript/ts-reset`
- * @link 
+ * @link
  */
 interface ReadonlyArray<T> {
-  filter(predicate: BooleanConstructor, thisArg?: any): NonFalsy<T>[];
+  filter(predicate: BooleanConstructor, thisArg?: any): NonFalsy<T>[]
 }
-
 
 interface JSON {
   /**
@@ -33,8 +30,5 @@ interface JSON {
    * @param reviver A function that transforms the results. This function is called for each member of the object.
    * If a member contains nested objects, the nested objects are transformed before the parent object is.
    */
-  parse(
-    text: string,
-    reviver?: (this: any, key: string, value: any) => any,
-  );
+  parse(text: string, reviver?: (this: any, key: string, value: any) => any)
 }
