@@ -34,21 +34,7 @@ export const optionsESM: Options = {
 export const optionsCJS: Options = {
   ...options,
   format: 'cjs',
-  dts: {
-    compilerOptions: {
-      target: 'ES5',
-      module: 'commonjs',
-      moduleResolution: 'node',
-    },
-  },
+  dts: true,
   banner: bannerCJS,
-  // esbuildOptions: (options) => {
-  //   options.footer = {
-  //     // This will ensure we can continue writing this plugin
-  //     // as a modern ECMA module, while still publishing this as a CommonJS
-  //     // library with a default export, as that's how ESLint expects plugins to look.
-  //     // @see https://github.com/evanw/esbuild/issues/1182#issuecomment-1011414271
-  //     js: 'module.exports = module.exports.default;',
-  //   }
-  // },
+  splitting: false,
 }
