@@ -1,5 +1,5 @@
 import { defineConfig } from 'tsup'
 
-import { optionsCJS, optionsESM } from './src/index'
+import { optionsCJS, optionsESM } from './src/index.ts'
 
 export default defineConfig([optionsCJS, optionsESM])
